# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Qoyod Vendor & Purchase Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendor-categories:
    get:
      operationId: listVendorCategories
      tags: [vendor-categories]
      summary: List all vendor categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendorCategory"
    post:
      operationId: createVendorCategory
      tags: [vendor-categories]
      summary: Create a vendor category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorCategory"
  /vendor-categories/{id}:
    get:
      operationId: getVendorCategory
      tags: [vendor-categories]
      summary: Get a vendor category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorCategory"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateVendorCategory
      tags: [vendor-categories]
      summary: Update a vendor category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorCategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorCategory"
    delete:
      operationId: deleteVendorCategory
      tags: [vendor-categories]
      summary: Delete a vendor category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Vendors ───────────────────────────────────────────────────────────────
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
...
  /vendors/import:
    post:
      operationId: importVendors
      tags: [vendors]
      summary: Import vendors from CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateVendor
      tags: [vendors]
      summary: Update a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete a vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Vendor Documents ──────────────────────────────────────────────────────
  /vendors/{vendorId}/documents:
    get:
      operationId: listVendorDocuments
      tags: [vendor-documents]
      summary: List documents for a vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VendorDocument"
    post:
      operationId: createVendorDocument
      tags: [vendor-documents]
      summary: Add a document to a vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDocument"
  /vendors/{vendorId}/documents/{docId}:
    put:
      operationId: updateVendorDocument
      tags: [vendor-documents]
      summary: Update a vendor document
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorDocumentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDocument"
    delete:
      operationId: deleteVendorDocument
      tags: [vendor-documents]
      summary: Delete a vendor document
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Purchase Requests ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ───────────────────────────────────────────────────────────────
...
    VendorCategory:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    VendorCategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    # ── Vendor ─────────────────────────────────────────────────────────────
    Vendor:
      type: object
      required: [id, companyName, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactPerson:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        bankAccountName:
          type: ["string", "null"]
        bankAccountNumber:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        swiftCode:
          type: ["string", "null"]
        bankBranch:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        categoryIds:
          type: array
          items:
            type: integer
        categories:
          type: array
          items:
            $ref: "#/components/schemas/VendorCategory"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/VendorDocument"
        totalSpent:
          type: ["number", "null"]
        transactionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    VendorInput:
      type: object
      required: [companyName]
      properties:
        companyName:
          type: string
          minLength: 1
        contactPerson:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        bankName:
          type: string
        bankAccountName:
          type: string
        bankAccountNumber:
          type: string
        iban:
          type: string
        swiftCode:
          type: string
        bankBranch:
          type: string
        notes:
          type: string
        categoryIds:
          type: array
          items:
            type: integer
...
    VendorImport:
      type: object
      required: [csvData]
      properties:
        csvData:
          type: string
    ImportResult:
      type: object
      required: [imported, errors]
      properties:
        imported:
          type: integer
        errors:
          type: array
          items:
            type: string

    # ── Vendor Document ────────────────────────────────────────────────────
    VendorDocument:
      type: object
      required: [id, vendorId, documentType, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        documentType:
          type: string
        documentNumber:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
          format: date
        fileUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    VendorDocumentInput:
      type: object
      required: [documentType]
      properties:
        documentType:
          type: string
          minLength: 1
        documentNumber:
          type: string
        expiryDate:
          type: string
          format: date
        fileUrl:
          type: string
        notes:
          type: string
...