# lib/api-zod/src/generated/types/vendor.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qoyod Vendor & Purchase Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { VendorCategory } from './vendorCategory';
import type { VendorDocument } from './vendorDocument';

export interface Vendor {
  id: number;
  companyName: string;
  /** @nullable */
  contactPerson?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankAccountName?: string | null;
  /** @nullable */
  bankAccountNumber?: string | null;
  /** @nullable */
  iban?: string | null;
  /** @nullable */
  swiftCode?: string | null;
  /** @nullable */
  bankBranch?: string | null;
  /** @nullable */
  notes?: string | null;
  categoryIds?: number[];
  categories?: VendorCategory[];
  documents?: VendorDocument[];
  /** @nullable */
  totalSpent?: number | null;
  /** @nullable */
  transactionCount?: number | null;
  createdAt: Date;
}