# lib/api-zod/src/generated/types/vendorDocument.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qoyod Vendor & Purchase Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface VendorDocument {
  id: number;
  vendorId: number;
  documentType: string;
  /** @nullable */
  documentNumber?: string | null;
  /** @nullable */
  expiryDate?: Date | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}