# lib/api-zod/src/generated/types/vendorImport.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Qoyod Vendor & Purchase Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface VendorImport {
  csvData: string;
}